+++ xsh.h
@@ -1,5 +1,7 @@
 #ifndef XSH_H
 #define XSH_H
+#define MAX_LINE 512
+
 #include <stdio.h>
 
 struct cmds {
@@ -8,6 +10,7 @@ struct cmds {
 
 void handle_sigint(int sig);
 void prompt();
+int cmdlogic(char *cmd);
 int interpret(char *file);
 
 #endif