+++ CHANGELOG.md
@@ -1,3 +1,4 @@
-Version is v1.3.5-stable
+Version is v1.4.0-stable
 
-1. Fixed a bug where echo didn't exclude " characters
+1. Moved all command logic in a separate function (int cmdlogic(char *cmd))
+2. Now, I can add proper return codes